requirements.txt
# PulpoAI - Dependencias Principales

# FastAPI y servidor
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Base de datos
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pgvector>=0.2.0

# Redis
redis>=5.0.0

# Weaviate
weaviate-client>=3.25.0

# HTTP requests
requests>=2.31.0

# Procesamiento de documentos
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
pytesseract>=0.3.10
python-magic>=0.4.27
ocrmypdf>=15.0.0

# Utilidades
python-dotenv>=1.0.0
python-multipart>=0.0.6
tiktoken>=0.5.0
anyio>=4.0.0
aiofiles>=23.0.0
prometheus-client>=0.19.0

# Logging
structlog>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Desarrollo
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

